Add files via upload
The era of mobile technology opens the windows to the android app. The websites are vanishing and the mobile phones are emerging. It’s the time to change from conventional websites to apps, which has become the part of our daily routine. In the scenario of the assignment, we were required to develop an android application on Online Job Portal.  So, we introduce ‘Job Vibe’ the android application.
Job Vibe acts as a portal between Job Seekers and Recruiters. It is helpful for fresher as well as young graduates, to get stated with their career. Using Job Vibe, one can search job matching his or her qualification and appear for the selection process once the application is being registered and accepted. It can be very helpful as it allows users of different profile to upload their educational details and search matching job according to their qualification. Also, every user can apply for multiple job at a time.
Job Vibe connects jobseekers and recruiters by accurately matching candidate profile to the relevant job openings through an advanced 2-way matching technology. While most online job portals focus on getting candidates the next job, Job Vibe focuses on the entire career growth of the candidate.